+++ packages/grafana-data/src/types/data.ts
@@ -1,3 +1,5 @@
+export type KeyValue<T = any> = { [s: string]: T };
+
 export enum LoadingState {
   NotStarted = 'NotStarted',
   Loading = 'Loading',
@@ -70,7 +72,6 @@ export interface AnnotationEvent {
   dashboardId?: number;
   panelId?: number;
   userId?: number;
-  regionId?: number;
   login?: string;
   email?: string;
   avatarUrl?: string;
@@ -81,4 +82,7 @@ export interface AnnotationEvent {
   text?: string;
   type?: string;
   tags?: string[];
+
+  // Currently used to merge annotations from alerts and dashboard
+  source?: any; // source.type === 'dashboard'
 }